Fixing nullpointer exception when removing groups
git-tfs-id: [https://tfsnpt.int.thomson.com:443/tfs/Cobalt_Collection/]$/Cobalt eBook Builder/Development;C799968

+++ ebookGeneratorCore/src/com/thomsonreuters/uscl/ereader/group/service/GroupServiceImpl.java
@@ -330,11 +330,13 @@ public class GroupServiceImpl implements GroupService {
 	
 	public void removeAllPreviousGroups(BookDefinition bookDefinition) throws Exception {
 		List<GroupDefinition> GroupDefinition = getGroups(bookDefinition);
-
-		for (GroupDefinition group : GroupDefinition) {
-			proviewClient.removeGroup(group.getGroupId(), group.getProviewGroupVersionString());
-			TimeUnit.SECONDS.sleep(2);
-			proviewClient.deleteGroup(group.getGroupId(), group.getProviewGroupVersionString());			
+		
+		if(GroupDefinition != null){
+			for (GroupDefinition group : GroupDefinition) {
+				proviewClient.removeGroup(group.getGroupId(), group.getProviewGroupVersionString());
+				TimeUnit.SECONDS.sleep(2);
+				proviewClient.deleteGroup(group.getGroupId(), group.getProviewGroupVersionString());			
+			}
 		}
 
 	}
